requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
django-redis==5.0.0